Restructure Purse and Wallet, Add Tests
