+++ crates/bcr-wallet-api/src/pocket/mod.rs
@@ -1,19 +1,66 @@
 use crate::{
-    MintConnector,
+    ClowderMintConnector,
     error::{Error, Result},
-    utils,
-    wallet::SafeMode,
+    wallet::types::SafeMode,
 };
+use async_trait::async_trait;
 use bcr_common::cashu::{
     self, Amount, CurrencyUnit, KeySet, KeySetInfo, ProofsMethods, amount::SplitTarget,
     nut00 as cdk00, nut01 as cdk01, nut03 as cdk03, nut07 as cdk07,
 };
+use bcr_wallet_core::{SendSync, types::SendSummary};
 use bcr_wallet_persistence::PocketRepository;
 use std::collections::{HashMap, HashSet};
 use uuid::Uuid;
 
 pub mod credit;
 pub mod debit;
+mod restore;
+
+/// trait that represents a single compartment in our wallet where we store proofs/tokens of the
+/// same currency emitted by the same mint
+#[async_trait]
+pub trait PocketApi: SendSync {
+    fn unit(&self) -> CurrencyUnit;
+    async fn balance(&self) -> Result<Amount>;
+    async fn receive_proofs(
+        &self,
+        client: &dyn ClowderMintConnector,
+        keysets_info: &[KeySetInfo],
+        proofs: Vec<cashu::Proof>,
+        safe_mode: SafeMode,
+    ) -> Result<(Amount, Vec<cashu::PublicKey>)>;
+    async fn prepare_send(&self, amount: Amount, infos: &[KeySetInfo]) -> Result<SendSummary>;
+    async fn send_proofs(
+        &self,
+        rid: Uuid,
+        keysets_info: &[KeySetInfo],
+        client: &dyn ClowderMintConnector,
+        safe_mode: SafeMode,
+    ) -> Result<HashMap<cashu::PublicKey, cashu::Proof>>;
+    async fn cleanup_local_proofs(
+        &self,
+        client: &dyn ClowderMintConnector,
+    ) -> Result<Vec<cashu::PublicKey>>;
+    async fn restore_local_proofs(
+        &self,
+        keysets_info: &[KeySetInfo],
+        client: &dyn ClowderMintConnector,
+    ) -> Result<usize>;
+    async fn delete_proofs(&self) -> Result<HashMap<cashu::Id, Vec<cashu::Proof>>>;
+    async fn return_proofs_to_send_for_offline_payment(
+        &self,
+        rid: Uuid,
+    ) -> Result<(Amount, HashMap<cashu::PublicKey, cashu::Proof>)>;
+    /// WARN: Only used for hacky offline pay by token - will be removed
+    async fn swap_to_unlocked_substitute_proofs(
+        &self,
+        proofs: Vec<cashu::Proof>,
+        keysets_info: &[KeySetInfo],
+        client: &dyn ClowderMintConnector,
+        send_amount: Amount,
+    ) -> Result<Vec<cashu::Proof>>;
+}
 
 ///////////////////////////////////////////// SendReference
 #[derive(Default, Clone)]
@@ -27,7 +74,7 @@ struct SendReference {
 // Removes Spent proofs from local DB
 async fn cleanup_local_proofs(
     db: &dyn PocketRepository,
-    client: &dyn MintConnector,
+    client: &dyn ClowderMintConnector,
 ) -> Result<Vec<cdk01::PublicKey>> {
     let ys = db.list_all().await?;
     let request = cdk07::CheckStateRequest { ys };
@@ -83,7 +130,7 @@ async fn swap(
     inputs: Vec<cdk00::Proof>,
     mut premints: HashMap<cashu::Id, cdk00::PreMintSecrets>,
     keysets: HashMap<cashu::Id, KeySet>,
-    client: &dyn MintConnector,
+    client: &dyn ClowderMintConnector,
     db: &dyn PocketRepository,
     safe_mode: SafeMode,
 ) -> Result<Amount> {
@@ -94,9 +141,9 @@ async fn swap(
         .flat_map(|premint| premint.blinded_messages())
         .collect();
     if let SafeMode::Enabled { expire, alpha_pk } = safe_mode {
-        let (fps, returned_blinds, exp, commitment) =
-            utils::compel_commitment(inputs.clone(), blinds.clone(), expire, alpha_pk, client)
-                .await?;
+        let (fps, returned_blinds, exp, commitment) = client
+            .post_commitment(inputs.clone(), blinds.clone(), expire, alpha_pk)
+            .await?;
         db.store_commitment(fps, returned_blinds, exp, commitment)
             .await?;
     }
@@ -144,7 +191,7 @@ async fn swap_proof_to_target(
     target_amount: Amount,
     seed: &[u8; 64],
     db: &dyn PocketRepository,
-    client: &dyn MintConnector,
+    client: &dyn ClowderMintConnector,
     safe_mode: SafeMode,
 ) -> Result<HashMap<cdk01::PublicKey, cdk00::Proof>> {
     let target = SplitTarget::Value(target_amount);
@@ -153,14 +200,9 @@ async fn swap_proof_to_target(
         cdk00::PreMintSecrets::from_seed(target_keyset.id, counter, seed, proof.amount, &target)?;
     let blinds = premint.blinded_messages();
     if let SafeMode::Enabled { expire, alpha_pk } = safe_mode {
-        let (fps, returned_blinds, exp, commitment) = utils::compel_commitment(
-            vec![proof.clone()],
-            blinds.clone(),
-            expire,
-            alpha_pk,
-            client,
-        )
-        .await?;
+        let (fps, returned_blinds, exp, commitment) = client
+            .post_commitment(vec![proof.clone()], blinds.clone(), expire, alpha_pk)
+            .await?;
         db.store_commitment(fps, returned_blinds, exp, commitment)
             .await?;
     }
@@ -230,7 +272,7 @@ async fn send_proofs(
     swap_proof: Option<(Amount, cdk01::PublicKey)>,
     seed: &[u8; 64],
     db: &dyn PocketRepository,
-    client: &dyn MintConnector,
+    client: &dyn ClowderMintConnector,
     target_swap_keysetid: Option<cashu::Id>,
     safe_mode: SafeMode,
 ) -> Result<HashMap<cdk01::PublicKey, cdk00::Proof>> {
@@ -295,7 +337,7 @@ async fn return_proofs_to_send_for_offline_payment(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utils::tests::MockMintConnector;
+    use crate::external::test_utils::tests::MockMintConnector;
     use bcr_common::{cashu::Proof, core::signature, core_tests};
     use bcr_wallet_persistence::MockPocketRepository;
     use cashu::nut02 as cdk02;
