+++ crates/bcr-wallet-core/Cargo.toml
@@ -6,6 +6,7 @@ license.workspace = true
 
 [dependencies]
 bcr-common = {workspace = true}
+bip39.workspace = true
 bitcoin.workspace = true
 chrono = {workspace = true}
 serde.workspace = true
